added modules/get-mock-data.js
